+++ uhppoted/doors.go
@@ -84,6 +84,19 @@ func (u *UHPPOTED) SetDoorControl(controller uint32, door uint8, mode types.Cont
 	return nil
 }
 
+func (u *UHPPOTED) SetDoorPasscodes(controller uint32, door uint8, passcodes ...uint32) error {
+	u.debug("set-door-passcodes", fmt.Sprintf("%v door:%v", controller, door))
+
+	response, err := u.UHPPOTE.SetSuperPasswords(controller, door, passcodes...)
+	if err != nil {
+		return fmt.Errorf("%w: %v", ErrInternalServerError, fmt.Errorf("%v  error setting door %v passcodes (%w)", controller, door, err))
+	}
+
+	u.debug("set-door-passcodes", fmt.Sprintf("response %+v", response))
+
+	return nil
+}
+
 func (u *UHPPOTED) OpenDoor(request OpenDoorRequest) (*OpenDoorResponse, error) {
 	u.debug("open-door", fmt.Sprintf("request  %+v", request))